Refactor: add send payment DTO
