+++ src/modules/payment/dto/send-payment.dto.ts
@@ -1,0 +1,12 @@
+import { IsString } from 'class-validator';
+
+export class SendPaymentDto {
+  @IsString()
+  sender: string;
+
+  @IsString()
+  amount: string;
+
+  @IsString()
+  token: string;
+}
