+++ src/modules/payment/payment.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { PaymentController } from './controllers/payment/payment.controller';
-import { PaymentService } from './services/payment/payment.service';
+import { PaymentController } from 'src/modules/payment/controllers/payment.controller';
+import { PaymentService } from 'src/modules/payment/services/payment.service';
 
 import { Wallet } from 'src/modules/wallet/entities/wallet.entity';
 import { ReceiverWallet } from 'src/modules/wallet/entities/receiver-wallet.entity';
